% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_dendrogram.R
\name{ggplot2_dendrogram}
\alias{ggplot2_dendrogram}
\title{Hierarchical clustering from a matrix using ggplot2 and ggdendro}
\usage{
ggplot2_dendrogram(data, title, samples, expGroups)
}
\arguments{
\item{data}{(compulsory) either a matrix, a DESeqTransform or a DESeqDataSet object.}

\item{title}{(optional) main title of the plot and the name of the file. Default: time stamp.}

\item{samples}{(optional) vector containing the sample names. Default: column names of data.}

\item{expGroups}{(optional) vector containing the name of the experimental expGroups (same order as in the columns of data). no default}
}
\value{
a PDF file.
}
\description{
Hierarchical clustering from a matrix using ggplot2 and ggdendro
}
\examples{
dat <- matrix(rnorm(1200), ncol=6)
ggplot2_dendrogram(dat, samples=1:6, expGroups=rep(c("A", "B"), 3), title="test")
}
