% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_density.R
\name{ggplot2_density}
\alias{ggplot2_density}
\title{Density plot from a data frame using ggplot2}
\usage{
ggplot2_density(data, title)
}
\arguments{
\item{data}{(compulsory) a data frame in which each column is a sample.}

\item{title}{(optional) main title of the plot and the name of the file. Default: time stamp.}
}
\value{
a PDF plot that contains one page per column in data.
}
\description{
Density plot from a data frame using ggplot2
}
\examples{
dat <- data.frame(A=rnorm(200), B=rnorm(200))
ggplot2_density(dat, title="test")
}
