% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_dotplot.R
\name{ggplot2_dotplot}
\alias{ggplot2_dotplot}
\title{Title}
\usage{
ggplot2_dotplot(data, genes, expGroups, samples, title)
}
\arguments{
\item{data}{(compulsory) matrix: each column is a sample. Row names should be gene names.}

\item{genes}{(compulsory) vector of gene names that you want to create dot plots for. They should be found in data's row names.}

\item{expGroups}{(optional) vector containing the name of the experimental expGroups (same order as in the columns of data). no default}

\item{samples}{(optional) vector containing the sample names. Default: column names of data.}

\item{title}{(optional) main title of the plot and the name of the file. Default: time stamp.}
}
\value{
a PDF file
}
\description{
Title
}
\examples{
dat <- matrix(rnorm(6), ncol=2, dimnames=list(c("A", "B", "C"), 1:2))
ggplot2_dotplot(dat, genes=c("A", "C"), samples=c("sample1", "sample2"))
}
