% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_pca.R
\name{ggplot2_pca}
\alias{ggplot2_pca}
\title{Title}
\usage{
ggplot2_pca(data, title, first_pc, second_pc, samples, groups)
}
\arguments{
\item{data}{(compulsory) a matrix.}

\item{title}{(optional) main title of the plot and the name of the file. Default: time stamp.}

\item{first_pc}{(optional) component plotted on the x-axis. Default: 1.}

\item{second_pc}{(optional) component plotted on the y-axis. Default: 2.}

\item{samples}{(optional) vector containing the sample names. Default: column names of data.}

\item{groups}{(optional) vector containing the name of the experimental groups (same order as in the columns of data). no default}
}
\description{
Title
}
\examples{
ggplot2_pca(dat, first_pc=1, second_pc=3, samples=1:6, groups=rep(c("A", "B"), 3), title="PCA")
}
